# Enhancing Performance by Adding Concurrency and Optional Features