@a8m 2023-08-04
Thanks for the contribution, @Gebes.

That’s not the first time we try this. See: https://github.com/ent/ent/pull/3242#issuecomment-1382145601