@Gebes 2023-08-04
Oh, I see. It is exactly the same change as I made, without the configuration and concurrent package execution. Since for the almost exact same amount of entities, my CPU takes 0.3s for the package load, 0.3s-0.6s for the generation of all templates and 8s for the formatting, this is a easy fix which should be approved.

The problem with goimports is not that we need to make a PR to optimize goimports, but its that parsing the syntax of the files takes a longer amount of time for each entity adding a dozen of more files, simple concurrent better erradicates this issue.